+++ lib/core/config.rb
@@ -34,6 +34,10 @@ class Config # rubocop:disable Metrics/ClassLength
     @app ||= load_app_from_options || load_app_from_env
   end
 
+  def app_prefix
+    current&.fetch(:name)
+  end
+
   def location
     @location ||= load_location_from_options || load_location_from_env || load_location_from_file
   end
@@ -111,8 +115,12 @@ class Config # rubocop:disable Metrics/ClassLength
 
   def find_app_config(app_name1)
     @app_configs ||= {}
-    @app_configs[app_name1] ||= apps.find do |app_name2, app_config|
-                                  app_matches?(app_name1, app_name2, app_config)
+
+    @app_configs[app_name1] ||= apps.filter_map do |app_name2, app_config|
+                                  next unless app_matches?(app_name1, app_name2, app_config)
+
+                                  app_config[:name] = app_name2
+                                  app_config
                                 end&.last
   end